refactor(plugins): split plugin loader responsibilities
